chore: restructured files for populations endpoint
