+++ graphql_service/resolver/population_model.py
@@ -1,0 +1,36 @@
+"""
+.. See the NOTICE file distributed with this work for additional information
+   regarding copyright ownership.
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+       http://www.apache.org/licenses/LICENSE-2.0
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+"""
+
+from pathlib import Path
+from typing import List
+import json
+
+from ariadne import ObjectType, QueryType
+from graphql import GraphQLResolveInfo
+
+
+QUERY_TYPE = QueryType()
+POPULATION_TYPE = ObjectType("Population")
+
+
+@QUERY_TYPE.field("populations")
+def resolve_populations(
+    _: None, info: GraphQLResolveInfo, genome_id: str = None
+) -> List:
+    population_metadata_file = (
+        Path(__file__).parents[2] / "common/file_model/population_metadata.json"
+    )
+    with open(population_metadata_file) as pop_file:
+        population_metadata = json.load(pop_file)
+    return population_metadata.get(genome_id, [])
